Add files via upload
Different users are there for the project; the users are presentin different Department at different places. By this, a request is inducedby one of the other users to interface with other user/users or sharingsome data with them. And The user Data Can be Stored in the ServerWhere they can access it any time There can be a condition where amessage is to be broadcasted to entire Office by a user. So, this paper isabout communication among divergent users present at different sites,
sharing this Different network. ONS Stands for the Office networkscenario.